Implemented the "notmedia" attribute for loading both scripts and styles.
Using this attribute allows you to load a CSS or JS file in situations where a particular media type or query is false, meaning you can now say, "load screen.js for desktop browsers but not smartphones" with the following syntax:

enhance({
  loadScripts: [
    {src: 'js/screen.js', media: 'screen', notmedia: 'screen and (max-device-width: 480px)'}
  ]
});

It's similar to the "not" keyword within media queries, except that this can be used safely in browsers that don't support media queries at all, as long as you're okay with false negatives (ie. notmedia: 'screen and (device-max-width: 480px)' will return false on Firefox because it's actually false, and return false on IE simply because it doesn't support media queries). 

In the end, a device that understands media queries will be sent device-specific assets and also be shielded from screen-specific assets. Devices that don't support media queries will get either basic.css, or screen enhancements if they happen to pass all tests but do not understand media queries (which is probably unlikely).